+++ src/main/java/de/skuzzle/ghpromexporter/version/Versions.java
@@ -1,0 +1,8 @@
+package de.skuzzle.ghpromexporter.version;
+
+public interface Versions {
+
+    String application();
+
+    String springBoot();
+}